fix: only delete nodes for registered nodeclaims

+++ pkg/controllers/nodeclaim/lifecycle/controller.go
@@ -174,28 +174,33 @@ func (c *Controller) finalize(ctx context.Context, nodeClaim *v1.NodeClaim) (rec
 		return reconcile.Result{}, fmt.Errorf("adding nodeclaim terminationGracePeriod annotation, %w", err)
 	}
 
-	nodes, err := nodeclaimutil.AllNodesForNodeClaim(ctx, c.kubeClient, nodeClaim)
-	if err != nil {
-		return reconcile.Result{}, err
-	}
-	for _, node := range nodes {
-		// If we still get the Node, but it's already marked as terminating, we don't need to call Delete again
-		if !node.DeletionTimestamp.IsZero() {
-			continue
-		}
-		// We delete nodes to trigger the node finalization and deletion flow
-		if err = c.kubeClient.Delete(ctx, node); client.IgnoreNotFound(err) != nil {
+	// Only delete Nodes if the NodeClaim has not been registered. Deleting Node's without the termination finalizer
+	// may result in leaked leases due to a kubelet bug until k8s 1.29. The Node should be garbage collected after the
+	// instance is terminated by CCM.
+	// Upstream Kubelet Fix: https://github.com/kubernetes/kubernetes/pull/119661
+	if nodeClaim.StatusConditions().Get(v1.ConditionTypeRegistered).IsTrue() {
+		nodes, err := nodeclaimutil.AllNodesForNodeClaim(ctx, c.kubeClient, nodeClaim)
+		if err != nil {
 			return reconcile.Result{}, err
 		}
+		for _, node := range nodes {
+			// If we still get the Node, but it's already marked as terminating, we don't need to call Delete again
+			if !node.DeletionTimestamp.IsZero() {
+				continue
+			}
+			// We delete nodes to trigger the node finalization and deletion flow
+			if err = c.kubeClient.Delete(ctx, node); client.IgnoreNotFound(err) != nil {
+				return reconcile.Result{}, err
+			}
+		}
+		// We wait until all the nodes associated with this nodeClaim have completed their deletion before triggering the finalization of the nodeClaim
+		if len(nodes) > 0 {
+			return reconcile.Result{}, nil
+		}
 	}
-	// We wait until all the nodes associated with this nodeClaim have completed their deletion before triggering the finalization of the nodeClaim
-	if len(nodes) > 0 {
-		return reconcile.Result{}, nil
-	}
-	var isInstanceTerminated bool
 	// We can expect ProviderID to be empty when there is a failure while launching the nodeClaim
 	if nodeClaim.Status.ProviderID != "" {
-		isInstanceTerminated, err = terminationutil.EnsureTerminated(ctx, c.kubeClient, nodeClaim, c.cloudProvider)
+		isInstanceTerminated, err := terminationutil.EnsureTerminated(ctx, c.kubeClient, nodeClaim, c.cloudProvider)
 		if err != nil {
 			// 404 = the nodeClaim no longer exists
 			if errors.IsNotFound(err) {
@@ -221,7 +226,7 @@ func (c *Controller) finalize(ctx context.Context, nodeClaim *v1.NodeClaim) (rec
 		// can cause races due to the fact that it fully replaces the list on a change
 		// Here, we are updating the finalizer list
 		// https://github.com/kubernetes/kubernetes/issues/111643#issuecomment-2016489732
-		if err = c.kubeClient.Patch(ctx, nodeClaim, client.MergeFromWithOptions(stored, client.MergeFromWithOptimisticLock{})); err != nil {
+		if err := c.kubeClient.Patch(ctx, nodeClaim, client.MergeFromWithOptions(stored, client.MergeFromWithOptimisticLock{})); err != nil {
 			if errors.IsConflict(err) {
 				return reconcile.Result{Requeue: true}, nil
 			}
